Java/src/test/java/org/black/lotus/kata/LongestSubstringWithoutRepeatingCharactersTest.java
package org.black.lotus.kata;

import org.junit.Test;

import java.util.LinkedHashMap;

import static org.junit.Assert.*;

public class LongestSubstringWithoutRepeatingCharactersTest {

    @Test
    public void should_find_longest_substring() {
        LinkedHashMap<Integer, Integer> map = new LinkedHashMap<>();
        map.put(1, 1);
        map.put(2, 2);

        LongestSubstringWithoutRepeatingCharacters solution = new LongestSubstringWithoutRepeatingCharacters();
        String s1 = "abcabcbb";
        assertEquals(3, solution.lengthOfLongestSubstring(s1));
        assertEquals(1, solution.lengthOfLongestSubstring("bbbbb"));
        assertEquals(3, solution.lengthOfLongestSubstring("pwwkew"));
        assertEquals(2, solution.lengthOfLongestSubstring("abba"));
    }
}